[Misc] Code changes
* The test module was prefixed with org instead of com
* Small changes
