+++ application-xpoll-api/pom.xml
@@ -65,6 +65,9 @@
       <plugin>
         <groupId>org.jvnet.jaxb2.maven2</groupId>
         <artifactId>maven-jaxb2-plugin</artifactId>
+        <configuration>
+          <generatePackage>com.xwiki.xpoll.rest.model.jaxb</generatePackage>
+        </configuration>
         <executions>
           <execution>
             <id>generate</id>
@@ -84,7 +87,7 @@
             <id>default</id>
             <configuration>
               <excludes>
-                com/xwiki/xpoll/model/jaxb/*.java
+                com/xwiki/xpoll/rest/model/jaxb/*.java
               </excludes>
             </configuration>
           </execution>
