+++ application-xpoll-test/application-xpoll-test-pageobjects/src/main/java/com/xwiki/xpoll/test/po/ActiveStatusViewPage.java
@@ -17,7 +17,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.xwiki.xpoll.test.po;
+package com.xwiki.xpoll.test.po;
 
 import java.util.ArrayList;
 import java.util.List;
