+++ application-xpoll-test/application-xpoll-test-pageobjects/src/main/java/com/xwiki/xpoll/test/po/XPollEditPage.java
@@ -17,7 +17,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.xwiki.xpoll.test.po;
+package com.xwiki.xpoll.test.po;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
